split contentmessage and contentasset
